requirements.txt
pywikibot
mwparserfromhell
requests
python-dateutil
